#236: Move adjoint tests inside test subdir; rename; combine conftests
