+++ Makefile
@@ -18,14 +18,12 @@ lint:
 test: lint
 	@echo "Running test suite..."
 	@cd test && make
-	@cd test_adjoint && make
 	@echo "PASS"
 
 coverage:
 	@echo "Generating coverage report..."
 	@python3 -m coverage erase
-	@python3 -m coverage run -a --source=goalie -m pytest -v test
-	@python3 -m coverage run -a --source=goalie -m pytest -v test_adjoint
+	@python3 -m coverage run --source=goalie -m pytest -v test
 	@python3 -m coverage html
 
 demo:
