+++ test/adjoint/examples/mesh-with-hole.geo
