+++ test/adjoint/examples/mesh-with-hole.msh
