+++ test/adjoint/examples/point_discharge2d.py
