+++ test/adjoint/examples/point_discharge3d.py
